Remove WebSocket support to focus on HTTP.
