+++ docs/external-links.md
@@ -1,3 +1,5 @@
 # External Links
 
-TODO.
+External links related to using Mangum.
+
+If you're interested in contributing to this page, please reference this [issue](https://github.com/jordaneremieff/mangum/issues/104) in a PR.
