+++ mangum/exceptions.py
@@ -10,13 +10,5 @@ class UnexpectedMessage(Exception):
     """Raise when an unexpected message type is received during an ASGI cycle."""
 
 
-class WebSocketClosed(Exception):
-    """Raise when an application closes the connection during the handshake."""
-
-
-class WebSocketError(Exception):
-    """Raise when an error occurs in a WebSocket event."""
-
-
 class ConfigurationError(Exception):
     """Raise when an error occurs parsing configuration."""
