+++ mkdocs.yml
@@ -16,10 +16,9 @@ nav:
     - Introduction: 'index.md'
     - Adapter: 'adapter.md'
     - HTTP: 'http.md'
-    - WebSockets: 'websockets.md'
     - Lifespan: 'lifespan.md'
     - ASGI Frameworks: 'asgi-frameworks.md'
-    # - External Links: 'external-links.md' todo
+    - External Links: 'external-links.md'
     - Release Notes: 'release-notes.md'
 
 markdown_extensions:
