+++ setup.py
@@ -10,16 +10,11 @@ setup(
     version="0.9.2",
     packages=find_packages(),
     license="MIT",
-    url="https://github.com/erm/mangum",
+    url="https://github.com/jordaneremieff/mangum",
     description="AWS Lambda & API Gateway support for ASGI",
     long_description=get_long_description(),
     python_requires=">=3.6",
     install_requires=["typing_extensions", "dataclasses; python_version < '3.7'"],
-    extras_require={
-        "aws": ["boto3"],
-        "postgresql": ["psycopg2", "boto3"],
-        "redis": ["redis", "boto3"],
-    },
     package_data={"mangum": ["py.typed"]},
     long_description_content_type="text/markdown",
     author="Jordan Eremieff",
