+++ tests/conftest.py
@@ -118,191 +118,5 @@ def mock_http_api_event(request):
     return event
 
 
-@pytest.fixture
-def mock_ws_connect_event() -> dict:
-    return {
-        "headers": {
-            "Accept-Encoding": "gzip, deflate, br",
-            "Accept-Language": "en-US,en;q=0.9",
-            "Cache-Control": "no-cache",
-            "Host": "test.execute-api.ap-southeast-1.amazonaws.com",
-            "Origin": "https://test.execute-api.ap-southeast-1.amazonaws.com",
-            "Pragma": "no-cache",
-            "Sec-WebSocket-Extensions": "permessage-deflate; " "client_max_window_bits",
-            "Sec-WebSocket-Key": "bnfeqmh9SSPr5Sg9DvFIBw==",
-            "Sec-WebSocket-Version": "13",
-            "User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_5) "
-            "AppleWebKit/537.36 (KHTML, like Gecko) "
-            "Chrome/75.0.3770.100 Safari/537.36",
-            "X-Amzn-Trace-Id": "Root=1-5d465cb6-78ddcac1e21f89203d004a89",
-            "X-Forwarded-For": "192.168.100.1",
-            "X-Forwarded-Port": "443",
-            "X-Forwarded-Proto": "https",
-        },
-        "isBase64Encoded": False,
-        "multiValueHeaders": {
-            "Accept-Encoding": ["gzip, deflate, br"],
-            "Accept-Language": ["en-US,en;q=0.9"],
-            "Cache-Control": ["no-cache"],
-            "Host": ["test.execute-api.ap-southeast-1.amazonaws.com"],
-            "Origin": ["https://test.execute-api.ap-southeast-1.amazonaws.com"],
-            "Pragma": ["no-cache"],
-            "Sec-WebSocket-Extensions": [
-                "permessage-deflate; " "client_max_window_bits"
-            ],
-            "Sec-WebSocket-Key": ["bnfeqmh9SSPr5Sg9DvFIBw=="],
-            "Sec-WebSocket-Version": ["13"],
-            "User-Agent": [
-                "Mozilla/5.0 (Macintosh; Intel Mac OS X "
-                "10_14_5) AppleWebKit/537.36 (KHTML, "
-                "like Gecko) Chrome/75.0.3770.100 "
-                "Safari/537.36"
-            ],
-            "X-Amzn-Trace-Id": ["Root=1-5d465cb6-78ddcac1e21f89203d004a89"],
-            "X-Forwarded-For": ["192.168.100.1"],
-            "X-Forwarded-Port": ["443"],
-            "X-Forwarded-Proto": ["https"],
-        },
-        "requestContext": {
-            "apiId": "test",
-            "connectedAt": 1564892342293,
-            "connectionId": "d4NsecoByQ0CH-Q=",
-            "domainName": "test.execute-api.ap-southeast-1.amazonaws.com",
-            "eventType": "CONNECT",
-            "extendedRequestId": "d4NseGc4yQ0FsSA=",
-            "identity": {
-                "accessKey": None,
-                "accountId": None,
-                "caller": None,
-                "cognitoAuthenticationProvider": None,
-                "cognitoAuthenticationType": None,
-                "cognitoIdentityId": None,
-                "cognitoIdentityPoolId": None,
-                "principalOrgId": None,
-                "sourceIp": "192.168.100.1",
-                "user": None,
-                "userAgent": "Mozilla/5.0 (Macintosh; Intel "
-                "Mac OS X 10_14_5) "
-                "AppleWebKit/537.36 (KHTML, like "
-                "Gecko) Chrome/75.0.3770.100 "
-                "Safari/537.36",
-                "userArn": None,
-            },
-            "messageDirection": "IN",
-            "messageId": None,
-            "requestId": "d4NseGc4yQ0FsSA=",
-            "requestTime": "04/Aug/2019:04:19:02 +0000",
-            "requestTimeEpoch": 1564892342293,
-            "routeKey": "$connect",
-            "stage": "Prod",
-        },
-    }
-
-
-@pytest.fixture
-def mock_ws_send_event() -> dict:
-    return {
-        "body": '{"action": "sendmessage", "data": "Hello world"}',
-        "isBase64Encoded": False,
-        "requestContext": {
-            "apiId": "test",
-            "connectedAt": 1564984321285,
-            "connectionId": "d4NsecoByQ0CH-Q=",
-            "domainName": "test.execute-api.ap-southeast-1.amazonaws.com",
-            "eventType": "MESSAGE",
-            "extendedRequestId": "d7uRtFvnyQ0FYmw=",
-            "identity": {
-                "accessKey": None,
-                "accountId": None,
-                "caller": None,
-                "cognitoAuthenticationProvider": None,
-                "cognitoAuthenticationType": None,
-                "cognitoIdentityId": None,
-                "cognitoIdentityPoolId": None,
-                "principalOrgId": None,
-                "sourceIp": "192.168.100.1",
-                "user": None,
-                "userAgent": None,
-                "userArn": None,
-            },
-            "messageDirection": "IN",
-            "messageId": "d7uRtfaKSQ0CE4Q=",
-            "requestId": "d7uRtFvnyQ0FYmw=",
-            "requestTime": "05/Aug/2019:05:52:10 +0000",
-            "requestTimeEpoch": 1564984330952,
-            "routeKey": "sendmessage",
-            "stage": "Prod",
-        },
-    }
-
-
-@pytest.fixture
-def mock_ws_disconnect_event() -> dict:
-    return {
-        "headers": {
-            "Host": "test.execute-api.ap-southeast-1.amazonaws.com",
-            "x-api-key": "",
-            "x-restapi": "",
-        },
-        "isBase64Encoded": False,
-        "multiValueHeaders": {
-            "Host": ["test.execute-api.ap-southeast-1.amazonaws.com"],
-            "x-api-key": [""],
-            "x-restapi": [""],
-        },
-        "requestContext": {
-            "apiId": "test",
-            "connectedAt": 1565140098258,
-            "connectionId": "d4NsecoByQ0CH-Q=",
-            "domainName": "test.execute-api.ap-southeast-1.amazonaws.com",
-            "eventType": "DISCONNECT",
-            "extendedRequestId": "eBql1FJmSQ0FrjA=",
-            "identity": {
-                "accessKey": None,
-                "accountId": None,
-                "caller": None,
-                "cognitoAuthenticationProvider": None,
-                "cognitoAuthenticationType": None,
-                "cognitoIdentityId": None,
-                "cognitoIdentityPoolId": None,
-                "principalOrgId": None,
-                "sourceIp": "101.164.35.219",
-                "user": None,
-                "userAgent": "Mozilla/5.0 (Macintosh; Intel "
-                "Mac OS X 10_14_6) "
-                "AppleWebKit/537.36 (KHTML, like "
-                "Gecko) Chrome/75.0.3770.142 "
-                "Safari/537.36",
-                "userArn": None,
-            },
-            "messageDirection": "IN",
-            "messageId": None,
-            "requestId": "eBql1FJmSQ0FrjA=",
-            "requestTime": "07/Aug/2019:01:08:27 +0000",
-            "requestTimeEpoch": 1565140107779,
-            "routeKey": "$disconnect",
-            "stage": "Prod",
-        },
-    }
-
-
-@pytest.fixture
-def mock_websocket_app():
-    async def app(scope, receive, send):
-        if scope["type"] == "websocket":
-            while True:
-                message = await receive()
-                if message["type"] == "websocket.connect":
-                    await send({"type": "websocket.accept", "subprotocol": None})
-                elif message["type"] == "websocket.receive":
-                    await send({"type": "websocket.send", "text": "Hello world!"})
-                elif message["type"] == "websocket.disconnect":
-                    close_code = message.get("code", 1000)
-                    await send({"type": "websocket.close", "code": close_code})
-                    return
-
-    return app
-
-
 def pytest_generate_tests(metafunc):
     os.environ["AWS_REGION"] = "ap-southeast-1"
